backend/requirements.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
PyJWT>=2.8.0
pymssql>=2.2.11
pymysql>=1.1.0
python-dotenv>=1.0.0
Werkzeug>=3.0.1
openpyxl>=3.1.2
pandas>=2.2.0
gunicorn>=21.2.0
